+++ src/plugins/makeOriginalClass.js
@@ -12,10 +12,10 @@
  *
  * @example
  *
- * import makeOriginalClass from 'dumb-bem/makers/makeOriginalClass';
+ * import * as pluginOriginalClass from 'dumb-bem/plugins/makeOriginalClass'
  *
- * makeOriginalClass('block', { className: 'btn btn-lg btn-success' });
+ * pluginOriginalClass.maker('block', { className: 'btn btn-lg btn-success' })
  *   // => 'btn btn-lg btn-success'
  *
  */
-export default (_, { className }) => className
+export const maker = (_, { className }) => className